requirements.txt
python-dotenv
google-genai
fastapi 
uvicorn 
pydantic
gradio
agno 